test/alternator: strengthen tests for empty attribute values
In this series we refactors GSI and change the validation of values
set in attributes which are GSI keys. It's important to test that when
updating attributes that are *not* GSI keys - and are either base-table
keys or normal non-key attributes - the validation didn't change.
For example, empty strings are still not allowed in base-table key
attributes, but are allowed (since May 2020 in DynamoDB) in non-key
attributes.

We did have tests in this area, but this patch strengthens them -
adding a test for non-key attribute, and expanding the key-attribute
test to cover the UpdateItem and BatchWriteItem operations, not just
PutItem.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>