alternator: overhaul implementation GSIs
This patch changes the implementation of GSIs, the Alternator version of
materialized views:

* Before this patch, the GSI's key columns were made into real columns
  in the table's schema, and the materialized view used that column in
  the view's key.

* After this patch, the GSI's key columns usually (when they are not
  the base table's keys, and not any LSI's key) are left in the ":attrs"
  map, just like any other non-key column. We use a new type of computed
  column (added in the previous patch) to extract the desired element from
  this map.

This overhaul of the GSI implementation doesn't change anything in the
functionality (and the Alternator test suite tries very hard to ensure
that), but will finally allow us in a later patch to add a GSI to an
already-existing table. It will now be possible because the GSI will
be able to pick up existing data from inside the ":attrs" map where it
is stored, instead of requiring the data in the map to be moved to a
stand-alone column as the previous implementation needed.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>