materialized views: allow compute_value to return NULL
Before this patch, a MV computed column, column_computation::compute_value()
must always return a value. In this patch, we allow compute_value() to
return an optional; If the optional is disengaged, it means that a value
could not be computed for this view key column, and the view row should
not be inserted.

We need this for Alternator, which allows items with a missing value for
a GSI's key, which simply means the item will reach the base table but
not to GSI. This feature is analogous to the "IS NOT NULL" requirement
in CQL's materialized views.

Note that we had to make the return value of compute_value() an
optional<bytes> and couldn't just rely on the emptiness of the returned
bytes object, because in theory some computed columns in the future
*may* want return an empty string value for the view key so it cannot
be used a sign that the row should be skipped. By the way, although
such an empty-string view key is allowed in Scylla (since commit
16229959002cde23f031b4d3bba52061f0b76cb7), it's not used in Alternator -
because DynamoDB forbids empty strings as GSI keys, and refuses an attempt
to write one.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>