test/alternator: test GSIs with different key types
All of the tests in test/alternator/test_gsi.py use strings as the GSI's
keys. This tests a lot of GSI functionality, but we implicitly assumed that
our implementation used an already-correct and already-tested implementation
of key columns and MV, which if it works for one type, works for other types
as well.

This assumption no longer holds in the new GSI implementation were
proposing in this series - it has, in its "computed column" implementation,
different code for different types of GSI key attributes (the supported
types are "S"tring, "B"ytes, and "N"umber).

So in this patch we add tests for writing (and reading from) different
types of GSI key attributes. These tests showed their importance when
the first draft of this series failed them.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>