test/alternator: test combination of LSI and GSI
The series containing this patch reimplements GSIs to stop forcing their
key attribute to become a real column in the schema - and let it remains
a member of the map ":attrs" like all non-key attributes. But the series
did NOT change LSIs, and these still force their key to become a real
column.

What the test in this patch does is to verify that using the same
attribute as a key of *both* GSI and LSI on the same table works.
It's actually quite surprising that it does: After all, the LSI
should force the attribute to become a real column (to which base
reads and writes go), but the GSI uses a computed column which reads
from ":attrs", no? Well, it turns out that view.cc's
value_getter::operator() always had a surprising exception which
"rescues" this test and makes it pass: Before using a computed column,
this code checks if a base-table column with the same name exists, and
if it does, it is used instead of the computed column! It's not clear
why this logic was chosen, but it turns out to be really useful for
making the test in this test pass. And it's important that if we ever
change that unintuitive behavior, we will have this test as a regression
test.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>