alternator: create LSI before GSI to use real columns for LSI keys
Because an LSI can only be created together with the table, we can
(and do) make its range key attribute into a real column in the schema.
But if the same attribute is also used as a GSI key, we need to use the
same real column - and not a computed column - for this GSI.

Because the existing code created GSIs first and then LSIs, we needed
to switch this order to first create the LSIs (and remember which
attributes were used as LSI keys) and only then the GSIs. Most of
the change in this patch is straighforward code movement, except the
small change to tracking the seen LSI range keys and using it to
decide whether to use real columns or computed columns.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>