test/alternator: additional tests for UpdateTable add/delete GSI
We already have tests for UpdateTable adding or deleting a GSI from
an existing table, but this patch adds additional tests for this
feature that focus on the case that the table already has an *LSI*
with the same key column as the GSI's key column. Because our
implementation treats GSI key columns differently if they are also
LSI key columns (when they are an LSI key column they are real schema
columns, when they are not they are members of a map) - it's important
to test also that UpdateTable does this correctly (we already have
tests for CreateTable creating a GSI and an LSI on the same attribute).

The two new tests pass on Amazon DynamoDB, and xfail on Alternator
because the UpdateTable feature isn't yet implemented.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>