test/alternator: split UpdateTable tests from test_gsi.py
test/alternator/test_gsi.py has grown pretty large, and a large chunk of
it (400 lines of code) are tests for one feature - using UpdateTable to
add or delete a GSI on an existing table. Let's move these tests to a
separate file, test_gsi_updatetable.py. Soon we'll add even more tests
to it, as this feature has a lot of aspects and corner cases that need
testing.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>