test/alternator: add test for missing attribute in item in LSI
Test that when a table has a LSI, if the indexed attribute is missing, the
item is added to the base table but not the index.

We already have exactly the same test for GSI in test_gsi.py, but forgot
to do write the same test for LSI, which caused us to miss a bug in the
recent GSI and LSI reimplementation. We did have the same scenario in
test_streams.py::test_streams_updateitem_old_image_lsi_missing_column
but this was a Alternator Streams test and we should have a more basic
test for this in test_lsi.py.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>