test/alternator: fix wrong test for UpdateTable metrics
The test we had for counting Alternator operations metrics ran the
UpdateTable request without any parameters, which isn't actually a
valid call - as the test test_updatetable_empty demonstrates. We want
to print an error in this case, just as DynamoDB does.

So let's fix the test to use a valid UpdateTable request, one that
does the silly BillingMode='PAY_PER_REQUEST' (which is already the
current setting, so nothing is really changed, but it's still counted
as an operation in the metric).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>