test/cqlpy: add test for view updates with TTL
This patch adds a test for the case where a view key is a regular column
in the base, and this regular column is set a value with a key. The view
row that is created needs to fully disappear - including its row marker
and data - when that TTL is reached. It turns out we didn't have a cqlpy
test for this case (though we did in C++ tests), so let's add one.

We have a similar test for the same situation in secondary indexes
(which behind the scenes use materialized views), in

cassandra_tests/validation/entities/secondary_index_test.py::
testIndexOnRegularColumnInsertExpiringColumn

This test helped reproduce a bug in one of my incorrect refactoring
attempts for materialized-view update generation.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>