test/alternator: fix flaky test
The Alternator test test_cql_rbac.py::test_rbac_updatetable_gsi was flaky.

This test wanted to check that without proper permissions, the GSI delete
operation was not permitted. Because permissions are cached, the attempt is
retried expected to miss permissions after a short while - but this retry
loop is wrong for GSI deletion: For GSI deletion, if the first attempt
succeeded (because old permissions were still cached), the GSI will be
gone and the next attempts will report it doesn't exist - not that
permissions are lacking.

The solution is to do a retry loop on a different UpdateTable operation
(a silly setting of BillingMode), and when *that* fails on permission
checks it means the new permissions are knwon - and we can immediately
try the GSI deletion operation that we really wanted to try.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>